requirements.txt
Flask
requests
Flask-SQLAlchemy
python-dotenv

